net: server wait ending all writes when stopping.
issue #62

+++ libdariadb/net/server.h
@@ -8,8 +8,8 @@ namespace net {
 class Server {
 public:
   struct Param {
-    int port;
-    Param(int _port) { port = _port; }
+    unsigned short port;
+    Param(unsigned short _port) { port = _port; }
   };
   Server(const Param &p);
   ~Server();
